fix(db): 首次 PostgreSQL 运行时验证 —— 修掉 5 类只在 pg 上炸的缺陷
此前所有运行时冒烟都在 SQLite（默认 feature）上做，而三种方言里只有
PostgreSQL 严格校验占位符语法与列类型。本轮第一次把后端跑在真实
PostgreSQL 上（45 个 GET + 全部写路径），抓到的都是 SQLite 上永远看不到的真缺陷：

1. 未加方言分支的 `?` 占位符（pg 只认 `$1`）→ `syntax error at or near ","`
   影响 23 处：JT1078 围栏/路线 16 个端点、媒体检索、鉴权码、推流国标绑定、
   `platform/exit`、`alarm/handle`、审计日志 elapsed_ms。
   新增 `dyn_where::dialect_sql()`（`?`→`$n`，跳过引号内的 `?`），
   固定 SQL 统一走它，三种方言继续共用同一份文本。

2. 列宽不匹配：Rust 按 i64 解码而 pg 列是 integer →
   `Rust type i64 ... not compatible with SQL type INT4`。
   11 张表的 id 改为 bigserial + 启动期幂等迁移 `ensure_pg_column_types()`；
   `PlatformChannel.platform_id/device_channel_id` 收回 i32。

3. `plate_color` 类型与前端载荷形状双双不匹配：前端「新增/编辑 JT1078 终端」
   必然 422（颜色是 el-select 的数字、省域是 el-input 的字符串，而后端 DTO 相反），
   且 pg 下只要是 varchar 就让 /terminal/list 整表 500。
   三方言统一为 INTEGER + 迁移；DTO 用 opt_string_flexible 两种都收。

4. ZLM `openRtpServer` 此前省略 `port` → 真实 ZLM 回
   `-300 Required parameter missed: "port","stream_id"`，
   `/api/push/start`、对讲收流、部分 INVITE 的收流端口申请全部失败。
   改为无条件发送 port（None→0 自动分配），实测返回 port=30020。

5. sqlx 语句缓存按 SQL 文本命中且不校验参数类型 →
   `insufficient data left in message`：同文本 DELETE 一处绑 i32、一处绑 i64，
   导致 `/api/record/plan/{update,delete}` 500。关闭 pg 语句缓存（兜住全仓同类
   问题）+ 统一该处参数类型。

顺带：region/group 国标编码重复从唯一约束冲突 500 改为明确 400；
`IdQuery.phone` 增加 `phoneNumber` 别名（WVP 参数名）。

证据：postgres 冒烟连跑 3 次一致（仅剩 3 项预期项：CSV 非 JSON / 重复平台被拒 /
代理指向不存在的 RTSP 源被 ZLM 404）；ZLM 日志确认 `openRtpServer?...&port=0`；
cargo test 702 passed；playwright 66 passed（新增 e2e/tests/jtTerminal.spec.ts 4 例）。
新增 docs/DB_DIALECT_NOTES.md 固化三条多方言规则。

+++ src/db/platform_channel.rs
@@ -9,8 +9,13 @@ use super::Pool;
 #[derive(Debug, Clone, Serialize, FromRow)]
 pub struct PlatformChannel {
     pub id: i32,
-    pub platform_id: Option<i64>,
-    pub device_channel_id: Option<i64>,
+    // `gb_platform_channel.platform_id` / `device_channel_id` 在 postgres 里是
+    // `integer`（对应 `gb_platform.id` / `gb_device_channel.id`，两者 Rust 侧都是
+    // i32）。这里曾经写成 `Option<i64>`，sqlite（动态类型）和 mysql（宽进严出）
+    // 都不报错，只有 postgres 会在解码时抛
+    // `Rust type i64 is not compatible with SQL type INT4`。
+    pub platform_id: Option<i32>,
+    pub device_channel_id: Option<i32>,
     pub custom_device_id: Option<String>,
     pub custom_name: Option<String>,
     pub custom_manufacturer: Option<String>,
